Moving  actual order creation to a service object
