+++ src/orders/src/main/java/mushop/orders/controllers/OrdersController.java
@@ -1,157 +1,43 @@
 /**
- ** Copyright © 2020, Oracle and/or its affiliates. All rights reserved.
- ** Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
+ * * Copyright © 2020, Oracle and/or its affiliates. All rights reserved.
+ * * Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
  **/
-package  mushop.orders.controllers;
+package mushop.orders.controllers;
 
-import mushop.orders.config.OrdersConfigurationProperties;
-import mushop.orders.entities.*;
-import mushop.orders.repositories.CustomerOrderRepository;
+import mushop.orders.entities.CustomerOrder;
 import mushop.orders.resources.NewOrderResource;
-import mushop.orders.services.AsyncGetService;
-import mushop.orders.services.MessagingService;
-import mushop.orders.values.OrderUpdate;
-import mushop.orders.values.PaymentRequest;
-import mushop.orders.values.PaymentResponse;
+import mushop.orders.services.OrdersService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.data.rest.webmvc.RepositoryRestController;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
-
-import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.Calendar;
-import java.util.List;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 
 
 @RepositoryRestController
 public class OrdersController {
     private final Logger LOG = LoggerFactory.getLogger(getClass());
 
-    @Autowired
-    private OrdersConfigurationProperties config;
 
     @Autowired
-    private AsyncGetService asyncGetService;
-
-    @Autowired
-    private MessagingService messagingService;
-
-    @Autowired
-    private CustomerOrderRepository customerOrderRepository;
-
-    @Value(value = "${http.timeout:5}")
-    private long timeout;
+    private OrdersService ordersService;
 
     @ResponseStatus(HttpStatus.CREATED)
     @RequestMapping(path = "/orders", consumes = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.POST)
     public
     @ResponseBody
     CustomerOrder newOrder(@RequestBody NewOrderResource item) {
-        try {
-            if (item.address == null || item.customer == null || item.card == null || item.items == null) {
-                throw new InvalidOrderException("Invalid order request. Order requires customer, address, card and items.");
-            }
-            LOG.info("Creating order {}", item);
-            LOG.debug("Starting calls");
-            Future<Address> addressFuture = asyncGetService.getObject(item.address, new ParameterizedTypeReference<Address>() {
-            });
-            Future<Customer> customerFuture = asyncGetService.getObject(item.customer, new ParameterizedTypeReference<Customer>() {
-            });
-            Future<Card> cardFuture = asyncGetService.getObject(item.card, new ParameterizedTypeReference<Card>() {
-            });
-            Future<List<Item>> itemsFuture = asyncGetService.getDataList(item.items, new
-                    ParameterizedTypeReference<List<Item>>() {
-                    });
-            LOG.debug("End of calls.");
 
-            //Calculate total
-            float amount = calculateTotal(itemsFuture.get(timeout, TimeUnit.SECONDS));
-
-            // Call payment service to make sure they've paid
-            PaymentRequest paymentRequest = new PaymentRequest(
-                    addressFuture.get(timeout, TimeUnit.SECONDS),
-                    cardFuture.get(timeout, TimeUnit.SECONDS),
-                    customerFuture.get(timeout, TimeUnit.SECONDS),
-                    amount);
-            LOG.info("Sending payment request: " + paymentRequest);
-            Future<PaymentResponse> paymentFuture = asyncGetService.postResource(
-                    config.getPaymentUri(),
-                    paymentRequest,
-                    new ParameterizedTypeReference<PaymentResponse>() {
-                    });
-            PaymentResponse paymentResponse = paymentFuture.get(timeout, TimeUnit.SECONDS);
-            LOG.info("Received payment response: " + paymentResponse);
-            if (paymentResponse == null) {
-                throw new PaymentDeclinedException("Unable to parse authorisation packet");
-            }
-            if (!paymentResponse.isAuthorised()) {
-                throw new PaymentDeclinedException(paymentResponse.getMessage());
-            }
-
-            //Persist
-            CustomerOrder order = new CustomerOrder(
-                    null,
-                    customerFuture.get(timeout, TimeUnit.SECONDS),
-                    addressFuture.get(timeout, TimeUnit.SECONDS),
-                    cardFuture.get(timeout, TimeUnit.SECONDS),
-                    itemsFuture.get(timeout, TimeUnit.SECONDS),
-                    null,
-                    Calendar.getInstance().getTime(),
-                    amount);
-            LOG.debug("Received data: " + order.toString());
-
-            CustomerOrder savedOrder = customerOrderRepository.save(order);
-            LOG.debug("Saved order: " + savedOrder);
-            asyncGetService.deleteResource(getCartURI(item.items  ));
-            OrderUpdate update = new OrderUpdate(savedOrder.getId(), null);
-            messagingService.dispatchToFulfillment(update);
-            return savedOrder;
-        } catch (TimeoutException e) {
-            throw new IllegalStateException("Unable to create order due to timeout from one of the services.", e);
-        } catch (InterruptedException | IOException | ExecutionException e) {
-            throw new IllegalStateException("Unable to create order due to unspecified IO error.", e);
+        if (item.address == null || item.customer == null || item.card == null || item.items == null) {
+            throw new InvalidOrderException("Invalid order request. Order requires customer, address, card and items.");
         }
-    }
-
-
-//    TODO: Add link to shipping
-//    @RequestMapping(method = RequestMethod.GET, value = "/orders")
-//    public @ResponseBody
-//    ResponseEntity<?> getOrders() {
-//        List<CustomerOrder> customerOrders = customerOrderRepository.findAll();
-//
-//        Resources<CustomerOrder> resources = new Resources<>(customerOrders);
-//
-//        resources.forEach(r -> r);
-//
-//        resources.add(linkTo(methodOn(ShippingController.class, CustomerOrder.getShipment::ge)).withSelfRel());
-//
-//        // add other links as needed
-//
-//        return ResponseEntity.ok(resources);
-//    }
-
-    private float calculateTotal(List<Item> items) {
-        float amount = 0F;
-        float shipping = 4.99F;
-        amount += items.stream().mapToDouble(i -> i.getQuantity() * i.getUnitPrice()).sum();
-        amount += shipping;
-        return amount;
+        return ordersService.createNewOrder(item);
     }
 
     @ResponseStatus(value = HttpStatus.NOT_ACCEPTABLE)
-    public class PaymentDeclinedException extends IllegalStateException {
+    public static class PaymentDeclinedException extends IllegalStateException {
         public PaymentDeclinedException(String s) {
             super(s);
         }
@@ -164,15 +50,5 @@ public class OrdersController {
         }
     }
 
-    private URI getCartURI(URI itemUri){
-        String cartPath = itemUri.getPath();
-        String cart = cartPath.substring(0, cartPath.lastIndexOf("/items"));
-        URI cartUri = null;
-        try {
-            cartUri = new URI(itemUri.getScheme()+"://"+itemUri.getHost()+cart+(itemUri.getQuery()==null?"":itemUri.getQuery()));
-        } catch (URISyntaxException e) {
-            e.printStackTrace();
-        }
-        return cartUri;
-    }
+
 }
